requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
networkx==3.1
requests==2.31.0
aiohttp==3.9.3